% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_package.r
\name{dgpd}
\alias{dgpd}
\title{Density of the Generalized Pareto distribution (GPD)}
\usage{
dgpd(x, xi, mu = 0, beta = 1)
}
\arguments{
\item{x}{a (positive) vector}

\item{xi}{a number (the tail index)}

\item{mu}{a number (the lower bound)}

\item{beta}{a number (the scaling paramater, default = 1)}
}
\value{
the density of the Pareto 1 distribution at points \code{x}
}
\description{
Density of the Generalized Pareto distribution (GPD)
}
\examples{
dgpd(2, xi=1/1.5, 1, 1)
}
