% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_package.r
\name{dpareto1}
\alias{dpareto1}
\title{import HMisc
import knitr
importFrom("graphics", "abline", "legend", "lines", "par", "plot")
importFrom("stats", "integrate", "optim", "optimise", "runif", "var")
importFrom("utils", "install.packages")
Density of the Pareto 1 distribution}
\usage{
dpareto1(x, mu = 1, alpha = 1/xi, xi = 1/alpha)
}
\arguments{
\item{x}{a (positive) vector}

\item{mu}{a number (the lower bound, default is 1)}

\item{alpha}{a number (the power index)}

\item{xi}{a number (the inverse of \code{alpha}, corresponding to the tail index)}
}
\value{
the density of the Pareto 1 distribution at points \code{x}
}
\description{
import HMisc
import knitr
importFrom("graphics", "abline", "legend", "lines", "par", "plot")
importFrom("stats", "integrate", "optim", "optimise", "runif", "var")
importFrom("utils", "install.packages")
Density of the Pareto 1 distribution
}
\examples{
dpareto1(2, 1, alpha=1.5, xi=1/alpha)
}
\seealso{
[ppareto1()], [qpareto1()] and [rpareto1()]
}
