% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_package_v3_juin_2019.R
\name{EPD}
\alias{EPD}
\title{Fit the Extended Pareto distribution to a vector of observations, with weights}
\source{
adapted from \url{https://github.com/TReynkens/ReIns/blob/master/R/EPD.R} Tom Reynkens, ReIns package version 1.0.7
}
\usage{
EPD(data, weights = rep(1, length(data)), rho = -1, start = NULL,
  direct = TRUE, warnings = FALSE, ...)
}
\arguments{
\item{data}{numeric, vector of observations}

\item{weights}{numeric, vector of (positive) weights}

\item{rho}{numeric, parameter of Fraga Alves et al. (2003) estimate}

\item{start}{numeric, vector of length 2 containing the starting values for the optimisation}

\item{direct}{logical, indicating if the parameters are obtained by directly maximising the log-likelihood function}

\item{warnings}{logical indicating if possible warnings from the optimisation function are shown}
}
\value{
a list with \code{k} the number of observations used, \code{gamma} the vector of the corresponding estimates for the tail parameter of the EPD, \code{kappa} the vector of the corresponding MLE estimates for the kappa parameter of the EPD and \code{tau} the vector of the corresponding estimates for the second order tail index parameter of the EPD using Hill estimates and values for \code{rho}
}
\description{
Fit the Extended Pareto distribution to a vector of observations, with weights
}
\examples{
set.seed(123)
x <- rpareto1(100,1,.5)
w <- rgamma(100,10,10)
EPD(data=x, weights=w)
}
\references{
Charpentier & Flachaire (2019) \emph{Pareto Models for Top Incomes } \href{https://hal.archives-ouvertes.fr/hal-02145024}{hal-02145024}
}
