% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_package_v4_juin_2019.R
\name{EPDfit}
\alias{EPDfit}
\title{Fit the Extended Pareto distribution to a vector of observations, with weights, using maximum likelihood estimation}
\source{
adapted from \url{https://github.com/TReynkens/ReIns/blob/master/R/EPD.R} Tom Reynkens, \code{ReIns} package version 1.0.7
}
\usage{
EPDfit(data, tau, start = c(0.1, 1), warnings = FALSE,
  weights = rep(1, length(data)))
}
\arguments{
\item{data}{numeric, vector of observations}

\item{tau}{numeric, value for tau in the EPD distribution}

\item{start}{numeric, vector of length 2 containing the starting values for the optimisation (default are \code{c(.1,1})}

\item{warnings}{logical indicating if possible warnings from the optimisation function are shown}

\item{weights}{numeric, vector of (positive) weights}

\item{rho}{numeric, parameter of Fraga Alves et al. (2003) estimate}
}
\value{
a list with \code{gamma} the vector of the corresponding estimates for the tail parameter of the EPD, \code{kappa} the vector of the corresponding MLE estimates for the kappa parameter of the EPD, as well as \code{tau}
}
\description{
Fit the Extended Pareto distribution to a vector of observations, with weights, using maximum likelihood estimation
}
\examples{
set.seed(123)
x <- rpareto1(100, mu=1, alpha=.5)
w <- rgamma(100,10,10)
EPDfit(data=x, tau=-3.3, weights=w)
}
\references{
Charpentier & Flachaire (2019) \emph{Pareto Models for Top Incomes } \href{https://hal.archives-ouvertes.fr/hal-02145024}{hal-02145024}
}
