% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_package_v4_juin_2019.R
\name{MLE.pareto1}
\alias{MLE.pareto1}
\title{Maximum Likelihood estimation of the Pareto 1 distribution, with weights}
\usage{
MLE.pareto1(data, weights = rep(1, length(x)), threshold = min(x))
}
\arguments{
\item{data}{a vector of observations}

\item{weights}{a vector of weights (default = 1)}

\item{threhold}{the threshold parameter of the Pareto 1 distribution (\code{mu})}
}
\value{
a list with the index \code{alpha} and \code{k}, the number of observations above \code{threshold}
}
\description{
Maximum Likelihood estimation of the Pareto 1 distribution, with weights
}
\examples{
set.seed(123)
x <- rpareto1(100, 1, 1.5)
w <- rgamma(100, 10, 10)
estim <- MLE.pareto1(data=x, weights=w, threshold=1)
estim
}
\references{
Charpentier & Flachaire (2019) \emph{Pareto Models for Top Incomes } \href{https://hal.archives-ouvertes.fr/hal-02145024}{hal-02145024}
}
