% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_package.r
\name{Pareto_diagram}
\alias{Pareto_diagram}
\title{Pareto diagrams - Pareto 1, GPD and EPD}
\usage{
Pareto_diagram(data, p = 0.01, q = 0.1, viz = TRUE)
}
\arguments{
\item{data}{dataframe obtained from \code{\link{tidy_income}} function}

\item{p}{numeric, the probability level (default 0.01 for the top 1\%)}

\item{q}{numeric, the probability level to model a Pareto distribution (default 0.1)}

\item{viz}{logical \code{TRUE} to plot the estimates}
}
\value{
a table with estimations of top share and a graph
}
\description{
Pareto diagrams - Pareto 1, GPD and EPD
}
\examples{
url_1 <- "https://github.com/freakonometrics/TopIncome/raw/master/data_csv/dataframe_yw_1.csv"
df <- read.table(url_1,sep=";",header=TRUE)
data_1  <-  tidy_income(income = df$y, weights = df$w)
\dontrun{Pareto_diagram(data_1)}
}
