% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_package_v4_juin_2019.R
\name{pepd}
\alias{pepd}
\title{Cumulative Distribution Function of the Extended Pareto distribution}
\source{
\url{https://github.com/TReynkens/ReIns/blob/master/R/Distributions.R} Tom Reynkens, \code{ReIns} package version 1.0.7
}
\usage{
pepd(x, gamma, kappa, tau = -1, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{x}{numeric, (positive) vector}

\item{gamma}{numeric, (strictly positive) number (the tail index)}

\item{kappa}{numeric, number - must be larger than max{-1,1/tau}}

\item{tau}{numeric, (negative) number (default is -1)}

\item{log}{logical, indicating if logarithm of density should be returned}
}
\value{
the c.d.f. of the Extended Pareto distribution at points \code{x}
}
\description{
Cumulative Distribution Function of the Extended Pareto distribution
}
\examples{
pepd(2,.5,1,-1)
## Expected value of a EPD(gamma=.5,kappa=1,tau=-1)
integrate(function(x) 1-pepd(x,.5,1,-1),0,Inf)
}
