% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_package_v4_juin_2019.R
\name{qepd}
\alias{qepd}
\title{Quantile Function of the Extended Pareto distribution}
\source{
\url{https://github.com/TReynkens/ReIns/blob/master/R/Distributions.R} Tom Reynkens, \code{ReIns} package version 1.0.7
}
\usage{
qepd(p, gamma, kappa, tau = -1, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{p}{numeric, vector of probabilities (in the interval [0,1])}

\item{gamma}{numeric, (strictly positive) number (the tail index)}

\item{kappa}{numeric, number - must be larger than max{-1,1/tau}}

\item{tau}{numeric, (negative) number (default is -1)}

\item{log}{logical, indicating if logarithm of density should be returned}
}
\value{
the c.d.f. of the Extended Pareto distribution at points \code{x}
}
\description{
Quantile Function of the Extended Pareto distribution
}
\examples{
qepd(.5,.5,1,-1)
}
