% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_package.r
\name{qpareto1}
\alias{qpareto1}
\title{Quantile function of the Pareto 1 distribution}
\usage{
qpareto1(p, mu, alpha = 1/xi, xi = 1/alpha)
}
\arguments{
\item{p}{a vector of probabilities (with values in [0,1])}

\item{mu}{a number (the lower bound)}

\item{alpha}{a number (the power index)}

\item{xi}{a number (the inverse of \code{alpha}, corresponding to the tail index)}
}
\value{
the quantile function of the Pareto 1 distribution at points \code{p}
}
\description{
Quantile function of the Pareto 1 distribution
}
\examples{
qpareto1(.5, 1, alpha=1.5)
}
