% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_package_v3_juin_2019.R
\name{repd}
\alias{repd}
\title{Random Generation of the Extended Pareto distribution}
\source{
\url{https://github.com/TReynkens/ReIns/blob/master/R/Distributions.R} Tom Reynkens, \code{ReIns} package version 1.0.7
}
\usage{
repd(n, gamma, kappa, tau = -1, quasimc = FALSE)
}
\arguments{
\item{n}{integer, number of generations}

\item{gamma}{numeric, (strictly positive) number (the tail index)}

\item{kappa}{numeric, must be larger than max{-1,1/tau}}

\item{tau}{numeric, (negative) number (default is -1)}

\item{quasimc}{logical, indicating quasi-Monte Carlo techniques (Halton sequence) should be used}
}
\value{
a vector of \code{n} values generated from an Extended Pareto distribution
}
\description{
Random Generation of the Extended Pareto distribution
}
\examples{
set.seed(123)
repd(6, gamma=.5,kappa=1,tau=-1)
repd(6, gamma=.5,kappa=1,tau=-1,quasimc=TRUE)
## Mean of a EPD(gamma=.5,kappa=1,tau=-1) distribution
mean(repd(1e4, gamma=.5,kappa=1,tau=-1))
mean(repd(1e4, gamma=.5,kappa=1,tau=-1),quasimc=TRUE)
}
