% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_package.r
\name{rgpd}
\alias{rgpd}
\title{Random generation of the Generalized Pareto distribution (GPD)}
\usage{
rgpd(n, xi, mu = 0, beta = 1)
}
\arguments{
\item{n}{an integer}

\item{xi}{a number (the tail index)}

\item{mu}{a number (the lower bound)}

\item{beta}{a number (the scaling paramater, default = 1)}
}
\value{
generates \code{n} values of the Pareto 1 distribution at points \code{x}
}
\description{
Random generation of the Generalized Pareto distribution (GPD)
}
\examples{
rgpd(10, xi=1/1.5, 1, 1)
}
