% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_package.r
\name{rpareto1}
\alias{rpareto1}
\title{Random generation of the Pareto 1 distribution}
\usage{
rpareto1(n, mu, alpha = 1/xi, xi = 1/alpha)
}
\arguments{
\item{n}{an integer}

\item{mu}{a number (the lower bound)}

\item{alpha}{a number (the power index)}

\item{xi}{a number (the inverse of \code{alpha}, corresponding to the tail index)}
}
\value{
generates \code{n} values of the Pareto 1 distribution
}
\description{
Random generation of the Pareto 1 distribution
}
\examples{
set.seed(123)
rpareto1(6, 1, alpha=1.5)
}
