% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_package.r
\name{Table_Top_Share}
\alias{Table_Top_Share}
\title{Table of top shares (using three thresholds)}
\usage{
Table_Top_Share(data, p = 0.01, q = c(0.1, 0.05, 0.01), md = FALSE,
  verbose = FALSE)
}
\arguments{
\item{data}{dataframe obtained from \code{tidy_income} function}

\item{p}{numeric, the probability level (default 0.01 for the top 1\%)}

\item{q}{numeric (possibly a vector), the probability level to model a Pareto distribution (default \code{c(0.1, 0.05, 0.01)} for the top 10\%, 5\% and 1\%)}

\item{md}{logical \code{TRUE} to have the table in a markdown format (default is \code{FALSE})}
}
\description{
Table of top shares (using three thresholds)
}
\examples{
set.seed(123)
x <- rpareto1(200, 1, alpha=1.5)
w <- rgamma(200, 10, 10)
data_0  <-  tidy_income(income = x, weights = w)
Table_Top_Share(data_0,q=c(.1,.05))
url_1 <- "https://github.com/freakonometrics/TopIncome/raw/master/data_csv/dataframe_yw_1.csv"
df <- read.table(url_1,sep=";",header=TRUE)
data_1  <-  tidy_income(income = df$y, weights = df$w)
Table_Top_Share(data_1)
}
