% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_package.r
\name{tidy_income}
\alias{tidy_income}
\title{tidy_income}
\usage{
tidy_income(income, weights)
}
\arguments{
\item{income}{a vector of data (income or wealth)}

\item{weights}{a vector of weight (same length as \code{income}}
}
\value{
a dataframe with 4 columns, \code{y} the vector income (or wealth), \code{weights} the vector of weights, \code{Fw} the cumulated proportion of people (with weights) and \code{Fx} the cumulated proportion of people (without weights)
}
\description{
Convert income/wealth data into a datafrane that can be used with \code{TopShare} or \code{Pareto_diagram}
}
\examples{
url_1 <- "https://github.com/freakonometrics/TopIncome/raw/master/data_csv/dataframe_yw_1.csv"
df <- read.table(url_1,sep=";",header=TRUE)
data_1  <-  tidy_income(income = df$y, weights = df$w)
str(data_1)
}
\seealso{
\code{\link{Top_Share}}, \code{\link{Pareto_diagram}}, \code{\link{Table_Top_Share}}
}
