% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_package_v3_juin_2019.R
\name{Top_Incomes}
\alias{Top_Incomes}
\title{Top Income plot}
\usage{
Top_Incomes(data, weights = rep(1, length(data)), p = 0.01,
  thr = seq(0.85, 0.999, by = 0.001), tail_index = TRUE, ...)
}
\arguments{
\item{data}{numeric, a vector of observations}

\item{p}{numeric, probability level (default 0.01)}

\item{thr}{numeric, vector of probability levels to model a Pareto distribution (default is \code{seq(.85,.999,by=.001)} from 0.85 up to 0.999)}

\item{tail_index}{logical to plot the tail index (default \code{TRUE})}

\item{weight}{numeric, a vector of weights (default is equal weights)}
}
\value{
one or two graphs (depending on \code{tail==TRUE})
}
\description{
Top Income plot
}
\examples{
################ Top Incomes on Synthetic Data
url_1 <- "https://github.com/freakonometrics/TopIncomes/raw/master/dataframe_yw_1.csv"
df <- read.table(url_1,sep=";",header=TRUE)
\dontrun{Top_Incomes(data = df$y, weights = df$w)}
}
\references{
Charpentier & Flachaire (2019) \emph{Pareto Models for Top Incomes } \href{https://hal.archives-ouvertes.fr/hal-02145024}{hal-02145024}
}
