% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_package.r
\name{Top_Share}
\alias{Top_Share}
\title{Top_Share}
\usage{
Top_Share(data, p = 0.01, q = 0.1, method = "edf",
  epd.direct = TRUE)
}
\arguments{
\item{data}{a vector of observations, from \code{\link{tidy_income}}}

\item{p}{numeric, the probability level (default 0.01 for the top 1\%)}

\item{q}{numeric (possibly a vector), the probability level to model a Pareto distribution (default 0.1)}

\item{method}{is the distribution considered (default \code{"edf"} for the empirical distribution function, but can be \code{"pareto1"}, \code{"gpd"} or  \code{"epd"})}

\item{edp.direct}{logical (default \code{TRUE}) for the method use for EPD fit}
}
\value{
estimation of the share of income/wealth owned by the top 100p\% of the population, assuming that the top 100q\% of the distribution is Pareto distributed. The list contains elements for each value of \code{q} if it is a vector, with \code{index} the share, \code{alpha} the inverse of the tail index, \code{coef},  \code{share.index} the value of \code{p}, \code{share.pareto} the value of the threshold as a percentage (.1 for the 90\% quantile) and \code{threshold} for the numerical value.
}
\description{
This function estimates Top Share
}
\examples{
url_1 <- "https://github.com/freakonometrics/TopIncome/raw/master/data_csv/dataframe_yw_1.csv"
df <- read.table(url_1,sep=";",header=TRUE)
data_1  <-  tidy_income(income = df$y, weights = df$w)
Top_Share(data_1)
}
