% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_package.r
\name{depd}
\alias{depd}
\title{Density of the Extended Pareto distribution}
\source{
\url{https://github.com/TReynkens/ReIns/blob/master/R/EPD.R} Tom Reynkens, ReIns package version 1.0.7
}
\usage{
depd(x, gamma, kappa, tau = -1, log = FALSE)
}
\arguments{
\item{x}{a (positive) vector}

\item{gamma}{a (strictly positive) number (the tail index)}

\item{kappa}{a number - must be larger than max{-1,1/tau}}

\item{tau}{a (negative) number (default is -1)}

\item{log}{logical indicating if logarithm of density should be returned}
}
\value{
the density of the Extended Pareto distribution at points \code{x}
}
\description{
Density of the Extended Pareto distribution
}
\examples{
depd(2,.5,1,-1)
}
