% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_package_v4_juin_2019.R
\name{dpareto1}
\alias{dpareto1}
\title{Density of the Pareto 1 distribution}
\usage{
dpareto1(x, mu = 1, alpha = 1/xi, xi = 1/alpha)
}
\arguments{
\item{x}{numeric, (positive) vector}

\item{mu}{numeric, lower bound, default is 1}

\item{alpha}{numeric, tail index}

\item{xi}{numeric, inverse of \code{alpha}}
}
\value{
The density of the Pareto 1 distribution at points \code{x} \deqn{f(x)=alpha*mu^alpha/x^(alpha+1)} Expected value exists when \code{alpha} is larger than 1, and is \deqn{E[X]=alpha/(alpha-1)}
}
\description{
Density of the Pareto 1 distribution
}
\examples{
# density of a Pareto1(1,1.5) at point 2
dpareto1(2, 1, alpha=1.5)
0.265165
# expected value of a Pareto1(1,1.5)
integrate(function(x) x*dpareto1(x, 1, alpha=1.5),-Inf,Inf)
# theoretical expected value
1*(1.5)/(1.5-1)
}
\seealso{
\code{\link{ppareto1}}, \code{\link{qpareto1}} and \code{\link{rpareto1}}
}
