% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_package_v3_juin_2019.R
\name{MLE.gpd}
\alias{MLE.gpd}
\title{Maximum Likelihood estimation of the Generalized Pareto distribution, with weights}
\usage{
MLE.gpd(data, weights = rep(1, length(x)), threshold = NA,
  nextremes = NA, method = "ml", information = c("observed",
  "expected"), ...)
}
\arguments{
\item{data}{numeric, vector of observations}

\item{weights}{numeric, vector of weights (default = 1)}

\item{method}{character, method used for inference (\code{"ml"} for maximum likelihood)}

\item{information}{character (not used)}

\item{threhol}{numeric, threshold parameter of the Generalized Pareto distribution (\code{mu})}

\item{nextrmes}{numeric, number of largest values considered (integer)}
}
\value{
a list \code{xi} the tail index, \code{mu} the threshold, \code{"beta"} the scaling coefficient, and \code{k} the number of observations above \code{mu}
}
\description{
Maximum Likelihood estimation of the Generalized Pareto distribution, with weights
}
\examples{
set.seed(123)
x <- rpareto1(100, 1, 1.5)
w <- rgamma(100,10,10)
estim <- MLE.gpd(data=x, weights=w, threshold=1)
estim$par.ests
}
\references{
Charpentier & Flachaire (2019) \emph{Pareto Models for Top Incomes } \href{https://hal.archives-ouvertes.fr/hal-02145024}{hal-02145024}
}
