% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_package_v4_juin_2019.R
\name{Pareto_diagram}
\alias{Pareto_diagram}
\title{Pareto diagrams - Pareto 1, GPD and EPD}
\usage{
Pareto_diagram(data, weights = rep(1, length(data)), p = 0.01,
  q = 0.1, viz = TRUE)
}
\arguments{
\item{data}{numeric, a vector of observations}

\item{p}{numeric, the probability level (default 0.01 for the top 1\%)}

\item{q}{numeric, the probability level to model a Pareto distribution (default 0.1)}

\item{viz}{logical, \code{TRUE} to plot the estimates}

\item{weight}{numeric, a vector of weights (default is equal weights)}
}
\value{
a table with estimations of top share and a graph
}
\description{
Pareto diagrams - Pareto 1, GPD and EPD
}
\examples{
################ Pareto diagams on Synthetic Data
url_1 <- "https://github.com/freakonometrics/TopIncomes/raw/master/dataframe_yw_1.csv"
df <- read.table(url_1,sep=";",header=TRUE)
\dontrun{Pareto_diagram(data = df$y, weights = df$w)}
}
\references{
Charpentier & Flachaire (2019) \emph{Pareto Models for Top Incomes } \href{https://hal.archives-ouvertes.fr/hal-02145024}{hal-02145024}
}
