% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_package_v3_juin_2019.R
\name{pgpd}
\alias{pgpd}
\title{Cumulative distribution of the Generalized Pareto distribution (GPD)}
\usage{
pgpd(x, xi = 1/alpha, alpha = 1/xi, mu = 0, beta = 1)
}
\arguments{
\item{x}{numeric, (positive) vector}

\item{xi}{numeric, inverse of \code{alpha}}

\item{alpha}{numeric, tail index}

\item{mu}{numeric, lower bound, default is 0}

\item{beta}{numeric, scaling paramater, default = 1}
}
\value{
the c.d.f. of the Generalized Pareto distribution at points \code{x}, \deqn{F(x)=1-(1+xi*(x-mu)/beta)^(-1/xi)} (assuming \code{xi} non null). The expected value is then \deqn{E[X]=mu+beta/(1-xi)}
}
\description{
Cumulative distribution of the Generalized Pareto distribution (GPD)
}
\examples{
pgpd(2, xi=1/1.5, 1, 1)
## expected value of a GPD(1,1.5,0,1) 
integrate(function(x) 1-pgpd(x,xi=1/1.5,mu=0,beta=1),0,Inf)
## theoretical value
0+1/(1-1/1.5)
}
\seealso{
\code{\link{dgpd}}, and \code{\link{rgpd}}
}
