% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_package_v3_juin_2019.R
\name{ppareto1}
\alias{ppareto1}
\title{Cumulative distribution function of the Pareto 1 distribution}
\usage{
ppareto1(x, mu = 1, alpha = 1/xi, xi = 1/alpha)
}
\arguments{
\item{x}{numeric, (positive) vector}

\item{mu}{numeric, lower bound, default is 1}

\item{alpha}{numeric, tail index}

\item{xi}{numeric, inverse of \code{alpha}}
}
\value{
the c.d.f. of the Pareto 1 distribution at points \code{x} \deqn{F(x)=1-(mu/x)^(alpha)} Expected value exists when \code{alpha} is larger than 1, and is \deqn{E[X]=alpha/(alpha-1)}
}
\description{
Cumulative distribution function of the Pareto 1 distribution
}
\examples{
# probability to a Pareto1(1,1.5) to be smaller than 2
ppareto1(2, 1, 1.5)
0.6464466
# expected value of a Pareto1(1,1.5) 
integrate(function(x) (1-ppareto1(x, 1, alpha=1.5)),0,Inf)
# theoretical expected value
1*(1.5)/(1.5-1)
}
\seealso{
\code{\link{dpareto1}}, \code{\link{qpareto1}} and \code{\link{rpareto1}}
}
