% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_package_v3_juin_2019.R
\name{qpareto1}
\alias{qpareto1}
\title{Quantile function of the Pareto 1 distribution}
\usage{
qpareto1(p, mu, alpha = 1/xi, xi = 1/alpha)
}
\arguments{
\item{p}{numeric, (probabilities) vector}

\item{mu}{numeric, lower bound, default is 1}

\item{alpha}{numeric, tail index}

\item{xi}{numeric, inverse of \code{alpha}}
}
\value{
the quantile function of the Pareto 1 distribution at points \code{p}
}
\description{
Quantile function of the Pareto 1 distribution
}
\examples{
qpareto1(.5, 1, alpha=1.5)
1.587401
# expected value of a Pareto1(1,1.5)
integrate(function(x) qpareto1(x, 1, alpha=1.5),0,1)
# theoretical expected value
1*(1.5)/(1.5-1)
}
\seealso{
\code{\link{dpareto1}}, \code{\link{ppareto1}} and \code{\link{rpareto1}}
}
