% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_package.r
\name{repd}
\alias{repd}
\title{Random Generation of the Extended Pareto distribution}
\source{
\url{https://github.com/TReynkens/ReIns/blob/master/R/Distributions.R} Tom Reynkens, ReIns package version 1.0.7
}
\usage{
repd(n, gamma, kappa, tau = -1)
}
\arguments{
\item{n}{integer, number of generations}

\item{gamma}{a (strictly positive) number (the tail index)}

\item{kappa}{a number - must be larger than max{-1,1/tau}}

\item{tau}{a (negative) number (default is -1)}

\item{log}{logical indicating if logarithm of density should be returned}
}
\value{
a vector of \code{n} values generated from an Extended Pareto distribution
}
\description{
Random Generation of the Extended Pareto distribution
}
\examples{
set.seed(123)
repd(6, gamma=.5,kappa=1,tau=-1)
}
