% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_package_v4_juin_2019.R
\name{ReturnEPD}
\alias{ReturnEPD}
\title{Large Return Period associated to the Extended Pareto distribution}
\source{
\url{https://github.com/TReynkens/ReIns/blob/master/R/Distributions.R} Tom Reynkens, ReIns package version 1.0.7
}
\usage{
ReturnEPD(data, q, gamma, kappa, tau, ...)
}
\arguments{
\item{data}{numeric, a vector of observations}

\item{q}{numeric, large quantile to estimate 1/P[X>q]}

\item{gamma}{numeric, vector of \code{n-1} estimates for the EVD obtained from [EPD]}

\item{kappa}{numeric, vector of \code{n-1} estimates for the EVD obtained from [EPD]}

\item{tau}{numeric, vector of \code{n-1} estimates for the EVD obtained from [EPD]}
}
\value{
a list with \code{k} the vector of the values of the tail parameter k, \code{R} the vector of the corresponding return period and \code{q} the used large quantile
}
\description{
Large Return Period associated to the Extended Pareto distribution
}
\examples{
set.seed(123)
x <- rpareto1(100, mu=1, alpha=1.5)
fit <- EPD(x)
ReturnEPD(data=x, q=.01, gamma=fit$gamma, kappa=fit$kappa, tau=fit$tau)
}
