% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_package_v3_juin_2019.R
\name{rgpd}
\alias{rgpd}
\title{Random generation of the Generalized Pareto distribution (GPD)}
\usage{
rgpd(n, xi = 1/alpha, alpha = 1/xi, mu = 0, beta = 1)
}
\arguments{
\item{n}{integer}

\item{xi}{numeric, inverse of \code{alpha}}

\item{alpha}{numeric, tail index}

\item{mu}{numeric, lower bound, default is 0}

\item{beta}{numeric, scaling paramater, default = 1}
}
\value{
generates \code{n} values of the Pareto 1 distribution at points \code{x}
}
\description{
Random generation of the Generalized Pareto distribution (GPD)
}
\examples{
rgpd(10, xi=1/1.5, 1, 1)
# expected value of a GPD(1,1.5,0,1)
mean(rgpd(1e6,alpha=1.5,mu=0,beta=1))
}
\seealso{
\code{\link{dgpd}}, and \code{\link{pgpd}}
}
