% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_package_v4_juin_2019.R
\name{Table_Top_Share}
\alias{Table_Top_Share}
\title{Table of top shares (using three thresholds)}
\usage{
Table_Top_Share(data, weights = rep(1, length(data)), p = 0.01,
  q = c(0.1, 0.05, 0.01), md = FALSE, verbose = FALSE)
}
\arguments{
\item{data}{numeric, a vector of observations}

\item{p}{numeric, the probability level (default 0.01 for the top 1\%)}

\item{q}{numeric (possibly a vector), the probability level to model a Pareto distribution (default \code{c(0.1, 0.05, 0.01)} for the top 10\%, 5\% and 1\%)}

\item{md}{logical \code{TRUE} to have the table in a markdown format (default is \code{FALSE})}

\item{weight}{numeric, a vector of weights (default is equal weights)}
}
\value{
top shares table (in a Markdown format if \code{md=TRUE})
}
\description{
Table of top shares (using three thresholds)
}
\examples{
################ Top Share on 200 Simulated Data
set.seed(123)
x <- rpareto1(200, 1, alpha=1.5)
w <- rgamma(200, 10, 10)
Table_Top_Share(data = x, weights = w, q = c(.1,.05))
################ Top Share on Synthetic Data
url_1 <- "https://github.com/freakonometrics/TopIncomes/raw/master/dataframe_yw_1.csv"
df <- read.table(url_1,sep=";",header=TRUE)
Table_Top_Share(data = df$y, weights = df$w)
}
\references{
Charpentier & Flachaire (2019) \emph{Pareto Models for Top Incomes } \href{https://hal.archives-ouvertes.fr/hal-02145024}{hal-02145024}
}
