% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_package_v3_juin_2019.R
\name{Top_Share}
\alias{Top_Share}
\title{Top_Share}
\usage{
Top_Share(data, weights = rep(1, length(data)), p = 0.01, q = 0.1,
  method = "edf", epd.direct = TRUE)
}
\arguments{
\item{data}{numeric, a vector of observations}

\item{p}{numeric, the probability level (default 0.01 for the top 1\%)}

\item{q}{numeric, (possibly a vector), the probability level to model a Pareto distribution (default 0.1)}

\item{method}{is the distribution considered (default \code{"edf"} for the empirical distribution function, but can be \code{"pareto1"}, \code{"gpd"} or  \code{"epd"})}

\item{weight}{numeric, a vector of weights (default is equal weights)}

\item{edp.direct}{logical (default \code{TRUE}) for the method used for EPD fit}
}
\value{
estimation of the share of income/wealth owned by the top 100p\% of the population, assuming that the top 100q\% of the distribution is Pareto distributed. The list contains elements for each value of \code{q} if it is a vector, with \code{index} the share, \code{alpha} the inverse of the tail index, \code{coef},  \code{share.index} the value of \code{p}, \code{share.pareto} the value of the threshold as a percentage (.1 for the 90\% quantile) and \code{threshold} for the numerical value.
}
\description{
This function estimates Top Share
}
\examples{
################ Top Share on Synthetic Data
url_1 <- "https://github.com/freakonometrics/TopIncomes/raw/master/dataframe_yw_1.csv"
df <- read.table(url_1,sep=";",header=TRUE)
## via the Empirical Distribution Function
Top_Share(data = df$y, weights = df$w,method="edf")$index
## via a Pareto 1 distribution (fitted on the top 100q\%)
Top_Share(data = df$y, weights = df$w,method="pareto1")$index
## via a Generalized Pareto distribution (fitted on the top 100q\%)
Top_Share(data = df$y, weights = df$w,method="gpd")$index
## via an Extended Pareto distribution (fitted on the top 100q\%)
Top_Share(data = df$y, weights = df$w,method="epd")$index
}
\references{
Charpentier & Flachaire (2019) \emph{Pareto Models for Top Incomes } \href{https://hal.archives-ouvertes.fr/hal-02145024}{hal-02145024}
}
